[O11Y-217] support seeking to specific row indexes (#50)
* [O11Y-217] support seeking to specific row indexes

* reset r.seek in rowGroupRowReader.WriteRowsTo

* revisit parquet.OffsetIndex API

* support seeking in row group row reader

* reduce interface conversions

* add NumRows/SeekToRow to parquet.Reader

* add tests for SeekToRow

+++ buffer_test.go
@@ -255,7 +255,7 @@ func testBuffer(t *testing.T, node parquet.Node, reader parquet.ValueReader, buf
 		}
 	}
 
-	if numRows := buffer.NumRows(); numRows != len(batch) {
+	if numRows := buffer.NumRows(); numRows != int64(len(batch)) {
 		t.Fatalf("number of rows mismatch: want=%d got=%d", len(batch), numRows)
 	}
 
@@ -274,7 +274,7 @@ func testBuffer(t *testing.T, node parquet.Node, reader parquet.ValueReader, buf
 
 	page := buffer.Column(0).(parquet.ColumnBuffer).Page()
 	numValues := page.NumValues()
-	if numValues != len(batch) {
+	if numValues != int64(len(batch)) {
 		t.Fatalf("number of values mistmatch: want=%d got=%d", len(batch), numValues)
 	}
